Multiple website updates (#73)
* MAPEX-170 Add VERIS ATT&CK Group mappings

This is a quick hack to get the VERIS group mappings (in JSON format)
onto the website. I filed another ticket to improve handling of groups
in the long run.

* MAPEX-141 Update docs for contributing

Added some text to the README explaining how to contribute

* MAPEX-163 Add methodology docs

* Fix build - move VERIS groups to a new legacy/ directory

This was failing because the VERIS groups file uses an outdated schema,
and everything under mappings/ or built under data/ needs to use the
proper schema. The solution is to create a new legacy/ directory that is
exempt from schema validation.